NexoraAI/backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
supabase>=2.0.0
upstash-redis>=0.15.0
google-generativeai>=0.5.0
httpx>=0.27.0
PyJWT>=2.8.0
xgboost>=2.0.0
scikit-learn>=1.4.0
joblib>=1.3.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
numpy>=1.26.0
pandas>=2.0.0
beautifulsoup4>=4.12.0
